Fix clippy warnings CI's -D warnings would have failed on
The release workflow runs clippy with -D warnings, which promotes every
lint to a hard error -- and the workspace was not actually clean under it.
Caught by literally running the same command CI does before trusting it,
rather than assuming the pipeline I wrote would pass.

- noise.rs and spike-h264-server: `x % n == 0` -> `x.is_multiple_of(n)`,
  the clippy-preferred form since it reads as one operation instead of two.
  Safe against the toolchain here: is_multiple_of stabilised in Rust 1.87,
  no rust-version is pinned anywhere in the workspace, and CI tracks stable.

- Four PipeWire spike/capture files: a stray `as u32` on a call already
  returning u32, left over across several PipeWire crate version bumps.
  No behaviour change, just a no-op cast the compiler now flags.

- palmtopd::modes::Mode: replaced a hand-written Default impl with
  #[derive(Default)] + #[default] on the Balanced variant -- what clippy
  suggests when a manual impl does nothing a derive couldn't.

Verified rather than assumed: `cargo clippy --workspace --all-targets --
-D warnings` now exits clean, all 35 tests still pass, and the daemon
restarts and runs normally after the capture.rs change.

+++ crates/spike-capture-encode/src/main.rs
@@ -221,11 +221,11 @@ fn run_capture(fd: std::os::fd::OwnedFd, node_id: u32, raw_path: PathBuf) -> Res
         id: pw::spa::param::ParamType::EnumFormat.as_raw(),
         properties: vec![
             Property::new(
-                FormatProperties::MediaType.as_raw() as u32,
+                FormatProperties::MediaType.as_raw(),
                 Value::Id(Id(MediaType::Video.as_raw())),
             ),
             Property::new(
-                FormatProperties::MediaSubtype.as_raw() as u32,
+                FormatProperties::MediaSubtype.as_raw(),
                 Value::Id(Id(MediaSubtype::Raw.as_raw())),
             ),
         ],
